+++ apps/manager-v2/src/app/manage-accounts/account.models.ts
@@ -1,4 +1,4 @@
-import { Account, AccountBlockingState, AccountConfig, AccountLimits, WelcomeUser } from '@nuclia/core';
+import { Account, AccountBlockingState, AccountConfig, AccountLimitsPatchPayload, WelcomeUser } from '@nuclia/core';
 import { Language, UserType } from '@flaps/core';
 
 export interface AccountSummary {
@@ -51,7 +51,7 @@ export interface AccountPatchPayload {
   max_dedicated_processors?: number;
   data?: object;
   blocking_state?: AccountBlockingState;
-  limits?: AccountLimits;
+  limits?: AccountLimitsPatchPayload;
 }
 
 export interface AccountKbList {